@@ -1,4 +1,4 @@
-class Photo < ActiveRecord::Base
+class Photo < ApplicationRecord
   belongs_to :user
   mount_uploader :image, ImageUploader
   validates :user_id, presence: true
