@@ -1,4 +1,4 @@
-class Rating < ActiveRecord::Base
+class Rating < ApplicationRecord
   belongs_to :portfolio
   validates :rating, presence: true;
   validates_inclusion_of :rating, :in => 1..10;
