@@ -1,4 +1,4 @@
-class Video < ActiveRecord::Base
+class Video < ApplicationRecord
   belongs_to :user
   validates :user_id, presence: true
   validates :category, presence: true
