@@ -24,7 +24,7 @@ module Canvas2
     config.i18n.default_locale = :ru
     # Do not swallow errors in after_commit/after_rollback callbacks.
     config.assets.initialize_on_precompile=false
-    config.active_record.raise_in_transactional_callbacks = true
+    # config.active_record.raise_in_transactional_callbacks = true
     config.assets.paths << Rails.root.join("app", "assets", "fonts")
     config.assets.paths << Rails.root.join("plugins/rating")
   end
